Adds conditional dashboard link to navbar
Enhances navigation by displaying a 'Dashboard Home' link in the navbar. This link is conditionally shown only when a user is authenticated and already within a dashboard route. It dynamically directs users to their specific dashboard based on their role (creator or buyer), streamlining access to their respective dashboard sections.